Refactor: adjust folder structure of PPMAPIModelLibrary
