+++ PPMAPIServiceLayer/InputDTOConverter/FinancialObjectFactory.cs
@@ -1,7 +1,7 @@
-﻿using PPMDTOModelLibrary.InputDTOs.FinancialInput;
-using PPMModelLibrary.Models.FinancialObjects;
-using PPMModelLibrary.Models.FinancialObjects.Transactions;
-using PPMModelLibrary.Models.FinancialObjects.ValueModifiers;
+﻿using PPMAPIModelLibrary.FinancialObjects;
+using PPMAPIModelLibrary.FinancialObjects.Transactions;
+using PPMAPIModelLibrary.FinancialObjects.ValueModifiers;
+using PPMDTOModelLibrary.InputDTOs.FinancialInput;
 
 namespace PPMAPIServiceLayer.InputDTOConverter
 {
