+++ PPMModelLibrary/FinancialObjects/ValueModifiers/IValueChange.cs
@@ -1,0 +1,7 @@
+﻿namespace PPMAPIModelLibrary.FinancialObjects.ValueModifiers
+{
+    internal interface IValueChange : IFinancialObject
+    {
+
+    }
+}
