--- PPMModelLibrary/Models/FinancialObjects/ValueModifiers/IValueChange.cs
@@ -1,7 +1,0 @@
-﻿namespace PPMModelLibrary.Models.FinancialObjects.ValueModifiers
-{
-    internal interface IValueChange : IFinancialObject
-    {
-        
-    }
-}
