+++ PPMModelLibrary/UtilityModels/IAddress.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace PPMModelLibrary.Models.UtilityModels
+namespace PPMAPIModelLibrary.UtilityModels
 {
     internal interface IAddress
     {
